--- src/components/common/modal/ModalPotal.js
@@ -1,7 +1,0 @@
-import { createPortal } from "react-dom";
-
-const ModalPortal = ({ children }) => {
-  return createPortal(<>{children}</>, document.getElementById("modal"));
-};
-
-export default ModalPortal;
