--- src/components/detail/TimeAgoDate.jsx
@@ -1,11 +1,0 @@
-import TimeAgo from "timeago-react";
-import * as timeAgo from "timeago.js";
-import ko from "timeago.js/lib/lang/ko";
-
-const TimeAgoDate = ({ createAt }) => {
-
-  
-  return ;
-};
-
-export default TimeAgoDate;
