+++ src/pages/CreateDiary.jsx
@@ -4,7 +4,7 @@ import { useSelector } from "react-redux";
 import styled from "styled-components";
 import { mainApi } from "../apis/axios";
 import NavigateBtn from "../components/common/NavigateBtn";
-import { DisplayDiv, StHeader } from "../UI/common";
+import { DisplayDiv, flex, StHeader } from "../UI/common";
 import { TiPencil } from "react-icons/ti";
 import soloDiaryBear from "../assets/images/soloDiaryBear.webp";
 import coupleDiaryBear from "../assets/images/coupleDiaryBear.webp";
@@ -89,10 +89,7 @@ const CreateDiaryBox = styled.section`
   width: 100%;
   height: calc(100% - 16.2rem);
   background-color: white;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+  ${flex("", "", "column")}
   input {
     margin-bottom: 2rem;
     width: 20.3rem;
@@ -110,10 +107,7 @@ const CreateDiaryBox = styled.section`
 `;
 
 const CreateLogoBear = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+  ${flex("", "", "", "column")}
   position: absolute;
   top: 10%;
   left: calc(50% - 15rem);
@@ -135,9 +129,7 @@ const Footer = styled.footer`
   width: 100%;
   height: 7.2rem;
   background-color: white;
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
+  ${flex("space-evenly", "")}
 `;
 
 const ColorPicker = styled.button`
