+++ src/pages/Invite.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { StContainer, StHeader, StSection } from "../UI/common";
+import { flex, StContainer, StHeader, StSection } from "../UI/common";
 import { BsSearch } from "react-icons/bs";
 import NavigateBtn from "../components/common/NavigateBtn";
 import { useRef, useState } from "react";
@@ -129,11 +129,9 @@ export default Invite;
 const StInviteSection = styled(StSection)``;
 
 const StSearchInputWrapper = styled.div`
-  display: flex;
+  ${flex("", "")}
   position: relative;
   gap: 1rem;
-  justify-content: center;
-  align-items: center;
   width: 100%;
   height: 5rem;
   & div {
@@ -175,9 +173,7 @@ const StSearchUserInfoWrapper = styled.div`
 `;
 
 const StSearchUserInfo = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
+  ${flex("space-between", "")}
   width: 100%;
   height: 5rem;
   padding: 0 1rem;
@@ -190,10 +186,7 @@ const StSearchUserInfo = styled.div`
     margin-right: 1rem;
   }
   div {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-start;
+    ${flex("", "flex-start", "column")}
     width: 100%;
     height: 100%;
     span {
